refactor resolutionScale for displayStruct to float

+++ Cylon.h
@@ -152,7 +152,7 @@ struct displayStruct
 	uint32_t	rotationPreference;
 	uint32_t	currentRotation;
 	uint32_t	nativeRotation;
-	uint32_t	resolutionScale;
+	float		resolutionScale;
 	float		logicalDPI;
 	float		rawDPIX;
 	float		rawDPIY;
